Se iguala estructura de modulo  usuarios como el resto.
