--- src/app/models/usuario.model.ts
@@ -1,7 +1,0 @@
-export interface Usuario {
-  id?: number;
-  nombre: string;
-  correo: string;
-  rol: string;
-  password?: string;
-}
