+++ src/app/usuarios/models/usuario.ts
@@ -1,0 +1,9 @@
+// Modelo de usuario
+export interface Usuario {
+  id: number;
+  nombre: string;
+  apellido: string;
+  email: string;
+  rol: string;
+  // Agrega más campos según sea necesario
+}
