--- src/app/usuarios/usuarios.service.ts
@@ -1,26 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { environment } from '../../environments/environment';
-import { Usuario } from '../models/usuario.model';
-
-@Injectable({ providedIn: 'root' })
-export class UsuariosService {
-  private apiUrl = environment.apiUrl + '/usuarios';
-
-  constructor(private http: HttpClient) {}
-
-  getUsuarios(): Observable<Usuario[]> {
-    return this.http.get<Usuario[]>(this.apiUrl);
-  }
-
-  getUsuarioPorId(id: number): Observable<Usuario> {
-    return this.http.get<Usuario>(`${this.apiUrl}/${id}`);
-  }
-
-  crearUsuario(usuario: Usuario): Observable<any> {
-    return this.http.post(this.apiUrl, usuario);
-  }
-
-  // Puedes agregar métodos para editar y eliminar usuarios si los endpoints existen
-}
